update: refactor application and add VIP table
