+++ config/feature_groups.py
@@ -1,0 +1,353 @@
+# Feature group definitions for PLS-DA analysis
+FEATURE_GROUPS_PLSDA = {
+    "index": {
+        "label": "Index Features",
+        "features": [
+            ("A", "index_A"),
+            ("Index B", "index_B"),
+            ("Norm A", "index_A_norm"),
+            ("Norm B", "index_B_norm"),
+            ("Max", "index_max"),
+            ("Min", "index_min"),
+            ("Average", "index_avg"),
+        ],
+    },
+    "atomic_weight": {
+        "label": "Atomic Weight",
+        "features": [
+            ("Weighted A+B", "atomic_weight_A+B_weighted"),
+            ("A/B", "atomic_weight_A/B"),
+            ("A-B", "atomic_weight_A-B"),
+        ],
+    },
+    "period_group": {
+        "label": "Period & Group",
+        "features": [
+            ("Period A", "period_A"),
+            ("Period B", "period_B"),
+            ("Group A", "group_A"),
+            ("Group B", "group_B"),
+            ("Group A-B", "group_A-B"),
+        ],
+    },
+    "mendeleev": {
+        "label": "Mendeleev Numbers",
+        "features": [
+            ("A", "Mendeleev_number_A"),
+            ("B", "Mendeleev_number_B"),
+            ("A-B", "Mendeleev_number_A-B"),
+        ],
+    },
+    "valence": {
+        "label": "Valence Electrons",
+        "features": [
+            ("A", "valencee_total_A"),
+            ("B", "valencee_total_B"),
+            ("A-B", "valencee_total_A-B"),
+            ("A+B", "valencee_total_A+B"),
+            ("Weighted A+B", "valencee_total_A+B_weighted"),
+            ("Norm Weighted A+B", "valencee_total_A+B_weighted_norm"),
+        ],
+    },
+    "unpaired": {
+        "label": "Unpaired Electrons",
+        "features": [
+            ("A", "unpaired_electrons_A"),
+            ("B", "unpaired_electrons_B"),
+            ("A-B", "unpaired_electrons_A-B"),
+            ("A+B", "unpaired_electrons_A+B"),
+            ("Weighted A+B", "unpaired_electrons_A+B_weighted"),
+            ("Norm Weighted A+B", "unpaired_electrons_A+B_weighted_norm"),
+        ],
+    },
+    "gilman": {
+        "label": "Gilman",
+        "features": [
+            ("A", "Gilman_A"),
+            ("B", "Gilman_B"),
+            ("A-B", "Gilman_A-B"),
+            ("A+B", "Gilman_A+B"),
+            ("Weighted A+B", "Gilman_A+B_weighted"),
+            ("Norm Weighted A+B", "Gilman_A+B_weighted_norm"),
+        ],
+    },
+    "z_eff": {
+        "label": "Effective Nuclear Charge",
+        "features": [
+            ("A", "Z_eff_A"),
+            ("B", "Z_eff_B"),
+            ("A-B", "Z_eff_A-B"),
+            ("A/B", "Z_eff_A/B"),
+            ("Max", "Z_eff_max"),
+            ("Min", "Z_eff_min"),
+            ("Avg", "Z_eff_avg"),
+            ("Norm Weighted A+B", "Z_eff_A+B_weighted_norm"),
+        ],
+    },
+    "ionization_energy": {
+        "label": "Ionization Energy",
+        "features": [
+            ("A", "ionization_energy_A"),
+            ("B", "ionization_energy_B"),
+            ("A-B", "ionization_energy_A-B"),
+            ("A/B", "ionization_energy_A/B"),
+            ("Max", "ionization_energy_max"),
+            ("Min", "ionization_energy_min"),
+            ("Avg", "ionization_energy_avg"),
+            ("Norm Weighted A+B", "ionization_energy_A+B_weighted_norm"),
+        ],
+    },
+    "coordination_number": {
+        "label": "Coordination Number",
+        "features": [
+            ("A", "coordination_number_A"),
+            ("B", "coordination_number_B"),
+            ("A-B", "coordination_number_A-B"),
+        ],
+    },
+    "ratio_closest": {
+        "label": "Ratio Closest",
+        "features": [
+            ("A", "ratio_closest_A"),
+            ("B", "ratio_closest_B"),
+            ("Max", "ratio_closest_max"),
+            ("Min", "ratio_closest_min"),
+            ("Avg", "ratio_closest_avg"),
+        ],
+    },
+    "polyhedron_distortion": {
+        "label": "Polyhedron Distortion",
+        "features": [
+            ("A", "polyhedron_distortion_A"),
+            ("B", "polyhedron_distortion_B"),
+            ("Max", "polyhedron_distortion_max"),
+            ("Min", "polyhedron_distortion_min"),
+            ("Avg", "polyhedron_distortion_avg"),
+        ],
+    },
+    "cif_radius": {
+        "label": "CIF Radius",
+        "features": [
+            ("A", "CIF_radius_A"),
+            ("B", "CIF_radius_B"),
+            ("A/B", "CIF_radius_A/B"),
+            ("A-B", "CIF_radius_A-B"),
+            ("Avg", "CIF_radius_avg"),
+            ("Norm Weighted A+B", "CIF_radius_A+B_weighted_norm"),
+        ],
+    },
+    "pauling_radius_cn12": {
+        "label": "Pauling Radius (CN12)",
+        "features": [
+            ("A", "Pauling_radius_CN12_A"),
+            ("B", "Pauling_radius_CN12_B"),
+            ("A-B", "Pauling_radius_CN12_A-B"),
+            ("A/B", "Pauling_radius_CN12_A/B"),
+            ("Avg", "Pauling_radius_CN12_avg"),
+            ("Norm Weighted A+B", "Pauling_radius_CN12_A+B_weighted_norm"),
+        ],
+    },
+    "pauling_en": {
+        "label": "Pauling Electronegativity",
+        "features": [
+            ("A", "Pauling_EN_A"),
+            ("B", "Pauling_EN_B"),
+            ("A-B", "Pauling_EN_A-B"),
+            ("A/B", "Pauling_EN_A/B"),
+            ("Max", "Pauling_EN_max"),
+            ("Min", "Pauling_EN_min"),
+            ("Avg", "Pauling_EN_avg"),
+            ("Norm Weighted A+B", "Pauling_EN_A+B_weighted_norm"),
+        ],
+    },
+    "martynov_batsanov_en": {
+        "label": "Martynov–Batsanov EN",
+        "features": [
+            ("A", "Martynov_Batsanov_EN_A"),
+            ("B", "Martynov_Batsanov_EN_B"),
+            ("A-B", "Martynov_Batsanov_EN_A-B"),
+            ("A/B", "Martynov_Batsanov_EN_A/B"),
+            ("Max", "Martynov_Batsanov_EN_max"),
+            ("Min", "Martynov_Batsanov_EN_min"),
+            ("Avg", "Martynov_Batsanov_EN_avg"),
+            ("Norm Weighted A+B", "Martynov_Batsanov_EN_A+B_weighted_norm"),
+        ],
+    },
+    "melting_point": {
+        "label": "Melting Point (K)",
+        "features": [
+            ("A", "melting_point_K_A"),
+            ("B", "melting_point_K_B"),
+            ("A-B", "melting_point_K_A-B"),
+            ("A/B", "melting_point_K_A/B"),
+            ("Max", "melting_point_K_max"),
+            ("Min", "melting_point_K_min"),
+            ("Avg", "melting_point_K_avg"),
+            ("Norm Weighted A+B", "melting_point_K_A+B_weighted_norm"),
+        ],
+    },
+    "density": {
+        "label": "Density",
+        "features": [
+            ("A", "density_A"),
+            ("B", "density_B"),
+            ("A-B", "density_A-B"),
+            ("A/B", "density_A/B"),
+            ("Max", "density_max"),
+            ("Min", "density_min"),
+            ("Avg", "density_avg"),
+            ("Norm Weighted A+B", "density_A+B_weighted_norm"),
+        ],
+    },
+    "specific_heat": {
+        "label": "Specific Heat",
+        "features": [
+            ("A", "specific_heat_A"),
+            ("B", "specific_heat_B"),
+            ("A-B", "specific_heat_A-B"),
+            ("A/B", "specific_heat_A/B"),
+            ("Max", "specific_heat_max"),
+            ("Min", "specific_heat_min"),
+            ("Avg", "specific_heat_avg"),
+            ("Norm Weighted A+B", "specific_heat_A+B_weighted_norm"),
+        ],
+    },
+    "cohesive_energy": {
+        "label": "Cohesive Energy",
+        "features": [
+            ("A", "cohesive_energy_A"),
+            ("B", "cohesive_energy_B"),
+            ("A-B", "cohesive_energy_A-B"),
+            ("A/B", "cohesive_energy_A/B"),
+            ("Max", "cohesive_energy_max"),
+            ("Min", "cohesive_energy_min"),
+            ("Avg", "cohesive_energy_avg"),
+            ("Norm Weighted A+B", "cohesive_energy_A+B_weighted_norm"),
+        ],
+    },
+    "bulk_modulus": {
+        "label": "Bulk Modulus",
+        "features": [
+            ("A", "bulk_modulus_A"),
+            ("B", "bulk_modulus_B"),
+            ("A-B", "bulk_modulus_A-B"),
+            ("A/B", "bulk_modulus_A/B"),
+            ("Max", "bulk_modulus_max"),
+            ("Min", "bulk_modulus_min"),
+            ("Avg", "bulk_modulus_avg"),
+            ("Norm Weighted A+B", "bulk_modulus_A+B_weighted_norm"),
+        ],
+    },
+}
+
+# Feature group definitions for PCA analysis  
+FEATURE_GROUPS_PCA = {
+    "basic_atomic_properties": {
+        "label": "Basic Atomic Properties",
+        "features": [
+            ("Atomic weight", "Atomic weight"),
+            ("Atomic number", "Atomic number"),
+            ("Period", "Period"),
+            ("Group", "Group"),
+            ("Families", "Families"),
+        ],
+    },
+    "atomic_and_ionic_radii": {
+        "label": "Atomic & Ionic Radii",
+        "features": [
+            ("Atomic radius calculated", "Atomic radius calculated"),
+            ("Covalent radius", "Covalent radius"),
+            ("Ionic radius", "Ionic radius"),
+            ("Effective ionic radius", "Effective ionic radius"),
+            ("Miracle radius", "Miracle radius"),
+            ("van der Waals radius", "van der Waals radius"),
+            ("Zunger radii sum", "Zunger radii sum"),
+            ("Crystal radius", "Crystal radius"),
+            ("Covalent CSD radius", "Covalent CSD radius"),
+            ("Slater radius", "Slater radius"),
+            ("Orbital radius", "Orbital radius"),
+            ("Polarizability", "polarizability, A^3"),
+            ("1st Bohr radius", "1st Bohr radius (a0)"),
+        ],
+    },
+    "electronegativity_and_electron_affinity": {
+        "label": "Electronegativity & Electron Affinity",
+        "features": [
+            ("Ionization energy", "Ionization energy (eV)"),
+            ("Electron affinity", "Electron affinity (ev)"),
+            ("Pauling EN", "Pauling EN"),
+            ("Martynov-Batsanov EN", "Martynov Batsanov EN"),
+            ("Mulliken EN", "Mulliken EN"),
+            ("Allred EN", "Allred EN"),
+            ("Allred-Rockow EN", "Allred Rockow EN"),
+            ("Nagle EN", "Nagle EN"),
+            ("Ghosh EN", "Ghosh EN"),
+        ],
+    },
+    "thermal_and_physical_properties": {
+        "label": "Thermal & Physical Properties",
+        "features": [
+            ("Melting point", "Melting point, K"),
+            ("Boiling point", "Boiling point, K"),
+            ("Density", "Density,  g/mL"),
+            ("Specific heat", "Specific heat, J/g K"),
+            ("Heat of fusion", "Heat of fusion,  kJ/mol"),
+            ("Heat of vaporization", "Heat of vaporization,  kJ/mol"),
+            ("Heat of atomization", "Heat of atomization,  kJ/mol"),
+            ("Thermal conductivity", "Thermal conductivity, W/m K"),
+            ("Cohesive energy", "Cohesive  energy"),
+            ("Bulk modulus", "Bulk modulus, GPa"),
+        ],
+    },
+    "electronic_structure": {
+        "label": "Electronic Structure",
+        "features": [
+            ("Mendeleev number", "Mendeleev number"),
+            ("Quantum number l", "quantum  number l"),
+            ("Valence s", "valence s"),
+            ("Valence p", "valence p"),
+            ("Valence d", "valence d"),
+            ("Valence f", "valence f"),
+            ("Unfilled s", "unfilled s"),
+            ("Unfilled p", "unfilled p"),
+            ("Unfilled d", "unfilled d"),
+            ("Onfilled f", "unfilled f"),
+            ("Valence electrons", "no. of  valence  electrons"),
+            ("Outer shell electrons", "outer shell electrons"),
+            ("Gilman valence electrons", "Gilman no. of valence electrons"),
+            ("Metallic valence", "Metallic  valence"),
+            ("Zeff", "Zeff"),
+        ],
+    },
+    "dft_lsd_and_lda_properties": {
+        "label": "DFT LSD & LDA Properties",
+        "features": [
+            ("DFT LDA Etot", "DFT LDA Etot"),
+            ("DFT LDA Ekin", "DFT LDA Ekin"),
+            ("DFT LDA Ecoul", "DFT LDA Ecoul"),
+            ("DFT LDA Eenuc", "DFT LDA Eenuc"),
+            ("DFT LDA Exc", "DFT LDA Exc"),
+            ("DFT LSD Etot", "DFT LSD Etot"),
+            ("DFT LSD Ekin", "DFT LSD Ekin"),
+            ("DFT LSD Ecoul", "DFT LSD Ecoul"),
+            ("DFT LSD Eenuc", "DFT LSD Eenuc"),
+            ("DFT LSD Exc", "DFT LSD Exc"),
+        ],
+    },
+    "dft_rlda_and_scrlda_properties": {
+        "label": "DFT RLDA & ScRLDA Properties",
+        "features": [
+            ("DFT RLDA Etot", "DFT RLDA Etot"),
+            ("DFT RLDA Ekin", "DFT RLDA Ekin"),
+            ("DFT RLDA Ecoul", "DFT RLDA Ecoul"),
+            ("DFT RLDA Eenuc", "DFT RLDA Eenuc"),
+            ("DFT RLDA Exc", "DFT RLDA Exc"),
+            ("DFT ScRLDA Etot", "DFT ScRLDA Etot"),
+            ("DFT ScRLDA Ekin", "DFT ScRLDA Ekin"),
+            ("DFT ScRLDA Ecoul", "DFT ScRLDA Ecoul"),
+            ("DFT ScRLDA Eenuc", "DFT ScRLDA Eenuc"),
+            ("DFT ScRLDA Exc", "DFT ScRLDA Exc"),
+        ],
+    },
+}
+
